Add Bootstrap navbar
Bootstrap has been used to create a navbar at the top of the page. The
links on the bar aren't currently working correctly, but the "Map" text
is set to active correctly when on the main page. Additionally, the
attribution of the map has been fixed to link to the OpenStreetMap
copyright notice.