Begin adding logic for dynamic GeoJSON
A start has been made on allowing GeoJSON data to be retrieved at runtime
instead of having to load it from a file. Storing all the GeoJSON files
for each country would use lots of space. Currently the Overpass API is
being used. Currently, it doesn't work as the JSON returned is in an OSM
form, and isn't a GeoJSON. There seems to be a relatively
straightforward way to convert the OSM JSON to a GeoJSON.