Add fetching GeoJSON data from Overpass API
The GeoJSON data needed to highlight countries on the map is now fetched
from the Overpass API instead of using local JSON files. Currently this
is only for countries that speak English, but French and Spanish will be
updated soon. Additionally, the GeoJSON returned has markers for each
country. These should be removed by ammending the url used to fetch the
data from the API.