Refactor language information and selection
Language information is now stored in maps. This means that long if
statements are no longer necessary to provide the correct information
for each language. In the future this could be moved into a separate
file. Additionally, the HTML for the language selection checkboxes has been
moved into a separate HTML file as it was getting long.